Add tenant creation on signup and onboarding flow fix
