+++ frontend/src/modules/auth/store/mutations.js
@@ -1,5 +1,7 @@
 import formbricks, { setupFormbricks } from '@/plugins/formbricks';
 import AuthCurrentTenant from '@/modules/auth/auth-current-tenant';
+import { isTrialExpired } from '@/utils/date';
+import config from '@/config';
 
 export default {
   CURRENT_USER_REFRESH_SUCCESS(state, payload) {
@@ -18,6 +20,12 @@ export default {
     state.currentTenant = AuthCurrentTenant.selectAndSaveOnStorageFor(
       payload.currentUser,
     );
+    if (isTrialExpired(state.currentTenant)) {
+      if (!window.location.href.includes('/onboard/plans')) {
+        window.location.href = `${config.frontendUrl.protocol}://${config.frontendUrl.host}/onboard/plans`;
+      }
+      return;
+    }
     state.loading = false;
 
     if (state.currentUser) {
@@ -111,6 +119,13 @@ export default {
     state.currentTenant = AuthCurrentTenant.selectAndSaveOnStorageFor(
       payload.currentUser,
     );
+    console.log(payload.currentTenant);
+    if (isTrialExpired(state.currentTenant)) {
+      if (!window.location.href.includes('/onboard/plans')) {
+        window.location.href = `${config.frontendUrl.protocol}://${config.frontendUrl.host}/onboard/plans`;
+      }
+      return;
+    }
     if (state.currentUser) {
       // initialize Formbricks
       setupFormbricks(state.currentUser);
