+++ frontend/src/modules/onboard/onboard-routes.js
@@ -3,6 +3,7 @@ import { PermissionChecker } from '@/modules/user/permission-checker';
 
 const OnboardPage = () => import('@/modules/onboard/pages/onboard-page.vue');
 const OnboardBookADemoPage = () => import('@/modules/onboard/pages/onboard-book-a-demo-page.vue');
+const OnboardPlansPage = () => import('@/modules/onboard/pages/onboard-plans-page.vue');
 
 export default [
   {
@@ -26,6 +27,15 @@ export default [
     meta: {
       auth: true,
       title: 'Onboarding',
+    },
+  },
+  {
+    name: 'onboardPlans',
+    path: '/onboard/plans',
+    component: OnboardPlansPage,
+    meta: {
+      auth: true,
+      title: 'Choose plan',
     },
   },
 ];
